Add directors and actors to movie insertion and display
Extended the admin panel and movie insertion API to support adding directors and actors (comma separated) when creating a movie. Updated the Gremlin queries to create and link director and actor vertices to movies. Enhanced the movie details page to display directors, actors, and improved the display of reviews and watched-by data in tables with additional details.